implementing replacement banks
