+++ cache.h
@@ -12,44 +12,38 @@
 
 #include "defaults.h"
 #include "types.h"
+#include "replacement_policies.h"	
 
 using namespace std;
 
+/* Cacheline */
 struct CacheLine {
 	uint tag;
 	bool valid;
 	bool dirty;
 	int number_data_blocks; //in B
 
-	/*	For replacement Policies based on Access Order
-		Replacement Element will have replacement_ordering = number_of_virtual_banks - 1;
-		FIFO, LRU
-	*/
-	int replacement_ordering;
-
-	/*	For replacement Policies based on Access count
-		LU
-	*/
-	double access_count;
-
+	/*Some pointer to actual data block ==> PENDING....*/
+	
 };
 
+/* CacheStats per cache */
 struct CacheStatistics {
 	uint hits;
 	uint misses;
 	uint access; //total memory references 
 	uint replacements; //total number of cache line replacements
 	uint write_backs; //total number of write backs
-	double bandwidth;
+	//double bandwidth;
 };
 
-/**
+/*
  * CacheLines is a collection of CacheLine objects, held in a vector.
  */
 typedef vector<CacheLine> CacheLines;
 
 
-/**
+/*
  * Describes a cache
  */
 class Cache {
@@ -70,10 +64,10 @@ class Cache {
 		Return true on hit*/
 		bool access(uint address, bool write);
 
-		/*Dump cache statistics on the output stream*/
+		/*Dump cache statistics on the Output Stream*/
 		void dump_stats(ostream& stream);
 
-		/*Dump cache settings on the output stream*/
+		/*Dump cache settings on the Output Stream*/
 		void dump_settings(ostream& stream);
 
 		/*Dump current cache state on the Output Stream*/
@@ -82,36 +76,34 @@ class Cache {
 		/*Run cache based on given input stream for #lines, -1 means complete run*/
 		bool run(istream& stream, int lines=-1); //run cache from stream.. can do cin as well
 
-		/*Set Virtual bank id and line in the bank to be replaced based on the replacement policy*/
+		/*Get back the virtual bank id to be replaced which is the index of the replacement_line*/
 		int get_replacement_line( CacheLines replacement_lines );
-
-		/*In case a line in the set needs to be replaced */
-		//int implement_replacement_policy(int index, int max_index, int policy_number); 
+	
 	
 	private:
-		/*Config*/
+
+		/*Configuration*/
 		uint size; //in B
 		int associativity; //should be to the power of 2
 		int number_virtual_banks; //keep track of virtual banks for implementing associativity 
 		int banks; //just use 1 for now
 		int number_cache_lines; //total number
-		int number_cache_sets;
+		int number_cache_sets; 
 		int number_data_blocks; //per cache line
-		//vector <cache_line> cache_lines;
 		vector<CacheLines> virtual_banks; //bank = collection of cache lines that belong to different sets
-		CacheStatistics stats;
-		Cache *lower_level;
-    	Cache *upper_level;
 		
 		/*Stats*/
-		cache_statistics stats; 
+		CacheStatistics stats;
 
 		/*Replacement Policies*/
-		//ReplacementPolicy replacement_policy;
-		int replacement_policy;
-		vector < vector <int> > max_replacement_ordering; //max_replacement_ordering[bank_id][set_index] = max. ordering
+		ReplacementPolicy replacement_policy;
+		//ReplacementStats *replacement_stats_table; 
 		
-		/*need hash functions for associativity later*/ 
+		/* Memory Hiearchy */
+		Cache *lower_level;
+    	Cache *upper_level;
+
+		/*need hash functions for associativity later: per bank*/ 
 };
 
 #endif
