+++ types.h
@@ -10,6 +10,7 @@ typedef unsigned int uint;
 enum ReplacementPolicy {
   RANDOM = 0,
   FIFO = 1,
+  LRU = 2,
 };
 
 #endif //MAIN_TYPES_H
